+++ pkg/infra/messages/messages.go
@@ -1,10 +1,11 @@
-package Messages
+package messages
 
 //Define message types to enable basic protocols, voting systems ...etc
-type Message struct {
+type Message interface {
 	// senderID
 	// content
 }
+
 type AckMessage struct {
 	msg Message
 	ack bool
