--- pkg/tower/tower.go
@@ -1,28 +1,0 @@
-package Tower
-
-import (
-	"fmt"
-
-	. "../agents/team1"
-)
-
-type Tower struct {
-	FoodOnPlatform  float64
-	FloorOfPlatform uint64
-	Agents          []BaseAgent
-}
-
-func (t Tower) TakeFood(a Agent, amt uint64) {
-	if foodPlatform_floor == a.floor {
-		if amt > t.PlatformFood {
-			a.hunger += t.PlatformFood
-			t.PlatformFood = 0
-			fmt.Println("Tried to take more food than exists, were given all the food")
-		} else {
-			a.hunger += amt
-			t.PlatformFood -= amt
-		}
-	} else {
-		fmt.Println("Platform not on your floor, no food for you")
-	}
-}
